Complain about UNRELEASED distribution in changelog, if changes file is signed. (Closes: #873520)
Emits the tag 'unreleased-changelog-distribution' with severity
wishlist.

Slurping the changes file is not great, but checks will ultimately be
wrapped in the their own process to free all memory they use. Changes
files are also small. The alternative is to shell out to grep, which
brings its own set of complications regarding result capture, error
handling and possible SIGCHILD conflicts with IO::Async.