Emit a wishlist tag for source format 1.0 packages. (Closes #884498)
Emits a gentle warning when packages use an older source format. The
tag has a somewhat visible priority.

This commit series is based on work by Chris Lamby, especially the
patch he submitted to the bug in 2017.

Changes in the test suite made it easier to implement this tag. Many
tests still use format 1.0, but the appearance there was no longer an
obstacle.  The selection menu for the interactive calibration of
expected tags now also offers an option to accept all changes. That
made it possible to adjust approximately 830 tests in one go.

The test suite will be upgraded to source format 3.0 separately.