Add newly appearing tag 'older-source-format' to tests that trigger it.
Accepts the new tag as expected for the tests in which it appears.

For four tests, three from the 'binaries-lfs-missing' family and one
called 'shared-libs-non-pic-i386' blind adjustments were made to avoid
failures on other architectures.  The tests do not build or run on the
author's platform amd64. Those modifications may not work.

We tried to find all tests that run only on other architectures, but
there could still be some that fail on other architectures. They would
not show until the build on that architecture fails as FTBFS, or as an
autopkgtest failure there.

Gbp-Dch: ignore