Specify explicit source format for tests using literal match strategy.
A small number of tests checks Lintian's exact literal output. For
those, the test runner does not offer an interactive adjustment of the
expected output based on test results. We can either add the tag for
'older-source-format' or explicitly specify an appropriate 3.0 format.
The latter path was taken.

Adds explicit, modern source formats for literal tests that require
it.

Made one blind adjustment for the test runtests-arch-i386, which does
not build or run on the author's platform amd64. That modification is
untested and may not work.

Gbp-Dch: ignore