In Lintian::Util::safe_qx, set $? to command's exit status.
Makes this IO::Async based routine more versatile by giving access to
the exit code. Corresponds more closely to the behavior of its
predecessor Lintian::Command::safe_qx, which was based on IPC::Run.

The routine no longer raises an exception on failure. Instead, it
returns the error as text, and sets $? to the exit status.

Gbp-Dch: ignore