Eliminate two remaining calls to IO::Async::Loop->really_new, which are rarely a good idea.
For some time, I received guidance to sometimes use ->really_new
instead of the plain ->new. Using the same loop led to confusion in
nested use as described in commit d4a90c21.

It seems the problem was resolved in IO::Async version 0.73 and
Lintian also currently undefines the loop singleton in nested usage
for backports.  The calls to ->really_new are no longer necessary.
This removes them.

Gbp-Dch: ignore