Add comment about silent failures when listing archives in collections.
There was some consideration about how to treat or prevent failures
for the 'ar' list command. It currently fails silently, and probably
creates empty entries further below.

If that ever becomes a problem, it is probably best addressed by

    next if $?;

or perhaps

    next unless length $output;

Other alternatives have been examined, such as:

    next if $file->file_info =~ /^ASCII text/;

or

    next unless $file->file_info =~ /^current ar archive/;

but they were ultimately general enough, and are not suitable.

Archives shipped in Debian packages do not have to be static
libraries, and are sometimes files in other formats. As a practical
matter, however, 'ar' fails for many of them:

    $ file /usr/lib/x86_64-linux-gnu/libmcheck.a
    /usr/lib/x86_64-linux-gnu/libmcheck.a: ELF 64-bit LSB relocatable, x86-64, version 1 (SYSV), not stripped
    $ ar t /usr/lib/x86_64-linux-gnu/libmcheck.a
    ar: /usr/lib/x86_64-linux-gnu/libmcheck.a: file format not recognized

or

    $ file /usr/share/go-1.12/src/go/internal/gcimporter/testdata/versions/test_go1.8_4.a
    /usr/share/go-1.12/src/go/internal/gcimporter/testdata/versions/test_go1.8_4.a: data
    $ ar t /usr/share/go-1.12/src/go/internal/gcimporter/testdata/versions/test_go1.8_4.a
    ar: /usr/share/go-1.12/src/go/internal/gcimporter/testdata/versions/test_go1.8_4.a: file format not recognized

The current behavior appears to work, so no action was taken. Some
background may be available under #934899.

Gbp-Dch: ignore