Fix test debconf/debconf-config-not-executable for new tag older-source-format.
As mentioned in commit dcc7ef82, there is a distinct possibility that
some test need this tag but do not run on the author's platform. I had
not considered backports. This test depends on an old version of dpkg.
There may be additional ones on other architectures.

This fixes the backports-stretch pipeline on jenkins.

Gbp-Dch: ignore