Remove tag 'source-field-does-not-match-pkg-name' from profile 'debian/ftp-master-auto-reject'.
The tag is no longer issued by Lintian. It was also not triggered for
any packages on lintian.d.o.

FTP Master was advised of this change via IRC and this merge request:

    https://salsa.debian.org/ftp-team/dak/merge_requests/154

The check was converted to a fatal program error. Users watching
Lintian's exit status and error output can probably still detect the
underlying condition.