Limit compat of one test for foreign binaries due to past debhelper regression.
Starting with version debhelper 12.5.2, objdump will no longer fail
silently in dh_makeshlibs for for compat level greater than 10. That
was not enforced until now.

Sets debhelper compat for the test binaries-from-other-arch to
level 10.

The binaries in the test appear to be dummies. Better binaries might
be helpful, but they would have to be cross-compiled.

This commit avoids the following error:

    objdump: debian/binaries-from-other-arch/usr/bin/elfobject: file truncated
    dh_makeshlibs: objdump -p debian/binaries-from-other-arch/usr/bin/elfobject returned exit code 1

Additional information may be available here:

    https://salsa.debian.org/debian/debhelper/commit/e1b3f5f78e6230a7ef161c08d1d204dad0777442#note_103885

Gbp-Dch: ignore