For one test, set minimum version of pkg-js-tools (>= 0.9.6~).
Debhelper 12.5.1 introduced a change with respect to add-on handling
for dh buildsystems. The pkg-js-tools package needed to be modified,
and so did debhelper. For details, please see Bug#935016.

Sets a minimum for of 0.9.6~ for pkg-js-tools (to accomodate
backports, as well).

That avoids the following error message:

    dh clean --with nodejs
    dh: unable to load addon nodejs: Debian/Debhelper/Sequence/nodejs.pm did not return a true value at (eval 10) line 1.
    BEGIN failed--compilation aborted at (eval 10) line 1.

Additional information may be available at:

    https://salsa.debian.org/debian/debhelper/commit/edf64efa348c91d631d090f36007fc1b872d02b3#note_103883

Gbp-Dch: ignore