Rename skeleton 'source' to 'source-native'.
For checks that only operate on source data, it is probably not
necessary to generate a full upload with 'dpkg-buildpackage'. In the
near future, all tests related to such checks may be converted to a
source skeleton. That should save some time.

The full builds (i.e. uploads) have different skeletons for native and
non-native tests. A non-native skeleton for source package will be
added shortly.

Renames the skeleton 'source' to 'source-native'.

Gbp-Dch: ignore