Add non-native source skeleton for tests.
For checks that only operate on source data, it is probably not
necessary to generate a full upload with 'dpkg-buildpackage'. In the
near future, all tests related to such checks may be converted to a
source skeleton. That should save some time.

The full builds (i.e. uploads) have different skeletons for native and
non-native tests.

Adds a non-native skeleton for source packages.

Gbp-Dch: ignore