In test suite, use a default 3.0 source format; add default version numbers.
Adds the version number 1.0 to the native skeletons for uploads and
sources as well as for the skeletons for deb and changes.

Adds the version numbers 1.0-1 to the non-native skeletons for uploads
and sources.

The version numbers seemed relatively common in the tests and could
probably be removed from many individual specifications, provided that
the code portions requiring the field to be present is modified.

Adds also adds the appropriate 3.0 source formats to the upload and
source skeletons.