Do not require version fields in individual test specifications.
The skeletons now provide default version numbers. It is no longer
necessary to specify a version number unless it is different from the
default. No longer requires the field to be present in each
specification.

Om addition, calculates parsed version portions only when a version is
present.

The test runner may in the future randomize version numbers for more
variety and broader coverage.

Finally, allows values from skeletons to override the overall
defaults. That was previously not possible due to the order in which
values were filled.

Gbp-Dch: ignore