Re-adds legitimate older-source-format tags for tests kept at source format 1.0.
In six tests, the tag older-source-format continues to be legitimate.
Four cruft-related tests were not converted to 3.0, as that seemed to
correspond more closely with their function and the author's intent.

Then there are two tracking tests for the generic setups offered by
dh_make in 2005 and 2008 that are generally not modified.

For these tests, the reversal of commit aa7108b9 is not appropriate.

Re-adds the tags that were removed by the revert.

Also includes one tag for debian-source-dir/debian-source-dir-general,
which is a test that removes the format supplied by the templates. It
will always test at an implied 1.0 format.

Gbp-Dch: ignore