Turn test python-script-uses-unversioned-python-in-shebang into a false positive for check python.
When the test suite was first sorted into checks, some general false
positives were not sorted. Those tests produce no output, and their
value is somewhat limited. More important, their value depends on what
the author said about them. A general false positive is worth much
less than a false positive specific to a particular tag.

When this test was first examined, it was thought to be quite general.
Upon review it appears useful as a false positive specific to the
check python. Adapts the test and places it there.

Changes the match strategy to 'tags' and converts the literal output
into the universal tag format. The test shows two classification tags.

The test was associated with the check 'python'.

Gbp-Dch: ignore