In tests, adjust expected tags for greater specificity; drop tags unrelated to check.
For tests that declare a check Lintian now runs only that check. That
reduces the set of expected tags to the tags issues by that particular
check. It makes the tests much more specific.

Removes all expected tags that are not related to the declared check.

The impact of this step is particularly noticable for the families of
tests that were duplicated as they were sorted into checks from the
previous test suite. Until now, that duplication had little value;
besides the ability to run only tests associated with a check, a
positive, the test suite just took longer to run.

With this commit, the expected tags for these families of tests began
to diverge severely. In fact, they no longer share any tags at all.

The most significant effect of this change, however, is running time.
In preliminary tests, the test suite is anywhere from 20% to 33%
faster than before.

Gbp-Dch: ignore