In Lintian, accept 'all' as a value for --check-part.
This changes avoid scripting the use of the option --check-part, which
is now used for almost all tests with a specific test.

For a small number of tests, this option is still used, but with
default value of 'all'. It made the design of the templating system
much easier.

This option still conflicts with other tag selection options even
though the value 'all' acts as if the option had not been specified at
all. It seemed more logical given the presence of the option.

Gbp-Dch: ignore