In Test::Lintian::Prepare, implement new templating order; also fill substitutions in substitution variables.
This changes the order in which values are loaded into the test case
structure. That leads to a more intuitive order in which values
override one another.

Now fills all values in fill-values.d and loads those values into the
list of substition variables to be used in templates. This is now a
powerful system that is generic enough to be used by any template that
needs it.

At some point, it may be desirable to switch the templating engine to
Template::Toolkit. It may also be better to use their delimiter.

Gbp-Dch: ignore