In Test::Lintian::Run, use new Lintian command line options and recognize Check: all.
The old way of specifying Lintian's command line options lead to
recursive behavior in fill-values.d. Introduces a new variable,
Lintian-Command-Line, as a solution.

The location of the executable was not specified as part of that
command line. That way, the runner has the option to use either the
unpacked or the installed version of Lintian. This feature is not
currently used but could be handy if package templating and building
are separated from the test suite.

The runner looks up whether the Check in a test exists in Lintian.
Tells it not to complain about a value of 'all'.

Gbp-Dch: ignore