Adapt test md5sums/md5sums-traversal to new templating system
This test used to specify which checks to run, but that is now
incompatible unless the default Lintian command line options are
overridden. Turns off the manual selection.

The selection of three checks before was unusual. From now on, the
test suite will run only the declared check md5sums.

One tag disappeared as a result.

Gbp-Dch: ignore