As testing default, run only declared Check; speeds up tests and makes them more specific.
For tests that declare a check Lintian now runs only that check. That
reduces the set of expected tags to the tags issues by that particular
check. It makes the tests much more specific. The most significant
effect, however, is running time.  In preliminary tests, the test
suite is anywhere from 20% to 33% faster than before.

Adds a default '--check-part' command line option to all invocations
of Lintian. Some tests that use tag selectors that conflict with that
option can and do override that value.

For Check:, adds a default value of 'all'. Lintian now accepts this
value.  It was necessary to avoid scripting the option '--check-part'.

For Fill-Values-Folder:, sets default for directory that holds
additional substitution variables to be added to desc. Templates can
add variables that contain substitutions themselves. The files in this
folder replace 'builder' and 'lintian-invocation' from before.  The
location of the folder can probably not be overridden because of the
way the files are loaded in Test::Lintian::Prepare.

The Sequence: value has not been used for some time and was already
removed from all individual tests. Also removes it from here.