Add script 'generate-qa-list-from-udd.pl' to generate 'qa-list.txt' from UDD database.
The script sends all package information to stdout. Debugging
information, if enabled with '-d', will go to stderr.

An automated invocation (in bash) would probably look similar to:

    TEMPFILE="$(mktemp)"
    "$LINTIAN_ROOT/reporting/generate-qa-list.pl" > "$TEMPFILE"
    mv "$TEMPFILE" "$HTML_DIR/qa-list.txt"

This script includes tags for all of 'unstable' and 'experimental',
but debug packages are not included. (I am trying to find them in
UDD.) That is why the counts produced by this script differ from the
original qa-list.txt.

This script aggregates the tags from all architectures and the two
releases through an ingenious row numbering and a DISTINCT query
(thanks to #postgresql). Like the original qa-list.txt, the resulting
counts do not necessarily correspond to any particular architecture or
release, but the results are probably the most useful for package
maintainers.

Gbp-Dch: ignore