Multi-faceted checks; invoke input-specific alternatives to ::run in checks.
Offers additional subroutines for particular types of inputs. Here, an
input is not a file on the command-line, but the type of processable,
i.e. source, binary, changes or udeb. That is also the kind of value
found in 'Type:' in a check's description.

This commit makes it possible to combine checks for source and binary
runs, such as 'changelog-file' and and 'source-changelog', in the same
file. Both can be copied together. Just change one ::run to ::binary
and the other to ::source.

The arguments to the new methods are the same as to ::run.

The routine ::run, if present, continues to be called, independent of
input.