Combine the checks changelog-file and source-changelog into a new check 'changelog'.
The division between these two checks was always awkward because both
deal with changelogs. This combines them into one called 'changelog'.

With two separate checks, it was super difficult to move tags from one
to the other. It required reassignment of tag descriptions,
duplication of tests, plus a recalibration of expected tags in both
the new and the remaining test.  Finally, the Test-Against: statements
had to be adjusted to include only tags belonging to the respective
check.

None of that is necessary when tags are moved from the 'binary' facet
to the 'source' facet of the new changelog test, as will happen
shortly when native heuristics are are restricted to sources, and
version parsing moves there.

All checks related to changelogs are now grouped conveniently in one
check.

Parts can still be moved out but that will represent meaningful areas
of examination (for example, for versions) and not simply expose
program internals.