Merge duplicate tests 'legacy-foo++' appearing in both checks about to be merged.
This test existed in both checks being merged. Both tests had the same
contents and functionality. The expected tags were merged and
resorted.  The Tags-Against: statements were combined.

This commit is part of an effort to merge the checks 'changelog-file'
and 'source-changelog' into a new check called 'changelog'.

Gbp-Dch: ignore