In tests, adjust expected tags for centralized version parsing.
Most of these changes are for tags that moved from binary to source.
Successful version parsing depends on determining whether a package is
native, and that can only be done reliably in a source package. As a
result, related tags are now source tags.

Also, the extra information for some tags, for example for epochs, was
modified to show the full version string (as opposed to just the
epoch). That may aid a maintainer in locating the offending changelog
entry more quickly.

Finally, the previous parsing methods offered more detailed advice
about why a version was unacceptable. The new parser offers fewer
tags, but they work better.

The two new tags are called 'malformed-debian-changelog-version' and
'odd-historical-debian-changelog-version'.

Gbp-Dch: ignore