In checks, allow new entry point 'always'; warn on outdated 'run'.
Checks are now encouraged to split their tags by binary, source, etch.
When that is not possible, a new entry point 'always' was created. It
functions like the old entry point 'run'. It is always called, even
when other components were run, which is reflected in the new name.

Warns when checks use the old entry point 'run'.

Not sure why the warning is only issued for the first offender.
Perhaps it has something to do with the class Test::Builder::Module.
This test would probably do fine using straight TAP.

Gbp-Dch: ignore