Associate tag 'malformed-debian-changelog-version' with check lintian.
This tag is now issued in Lintian::Collect::Source and thus no longer
associated with a traditional check. The tag is instead paired with
the placeholder check 'lintian'.

Centralized version parsing was instituted to improve accuracy. It
requires information about nativeness that is only available from
source packages.

Gbp-Dch: ignore