Add new tag 'odd-historical-debian-changelog-version'.
This tag is issued in check 'changelog' when the format of historical
changelog versions does not match expectations.

That can happen simply because there was a parsing error, but also
because a package changed from native to non-native (or vice versa).
The resulting tags will probably never change in the future (except
perhaps the tag output format), so they are easy to override.

The severity of the new tag is minor.

Gbp-Dch: ignore