Adapt test defaults to separate specifications for building and testing of test packages.
Splits the default values for 'desc' and 'fill-values' similar to how
they will be split for the individual tests. The build specifications
are now listed in 'fill-values', which refers to the filling of
templates.

Uses the new skeleton 'testing' in t/defaults/desc.

This commit is part of a series to allow adaptive and partial building
of test packages. It will ultimately solve Bug#935907.

Gbp-Dch: ignore