Re-focus test 'binaries-missing-lfs' on the tag 'binary-file-built-without-LFS-support'. (Closes #939639)
Starting in 2.20.0 the test suite invokes only the Check: listed in
'desc', if present. The test suite does not currently build packages
unrelated to host architecture, and was not properly adjusted on a
maintainer's amd64 platform.

The three similar tests exist because the original test had ties to
three checks.  It listed tags from 'binaries' and 'shared-libs', and
also declared that it was a false-positive against
wrong-section-according-to-package-name, which is part of 'files'.
Upon examination the test was meant for binaries. The other
appearances are spurious.

Removes the two copies of the test from the checks shared-libs and
fields and adjusts the expected tags in the remaining test to belong
only to the check binaries.

This change is untested due to the architecture mismatch.