Remove parts related to gzip timestamps from test 'files-general'. (Closes: #939673)
According to the changelog for gzip, these timestamps are no longer
issued starting in version 1.10:

* Noteworthy changes in release 1.10 (2018-12-29) [stable]

  Compressed gzip output no longer contains the current time as a
  timestamp when the input is not a regular file.  Instead, the output
  contains a null (zero) timestamp.  This makes gzip's behavior more
  reproducible when used as part of a pipeline.  (As a reminder, even
  regular files will use null timestamps after the year 2106, due to a
  limitation in the gzip format.)

Removes the components related to gzip timestamps from the test
files-general to avoid triggering an error with future versions of
gzip.

A separate test 'timestamp-in-piped-gzip' already has those parts and
is limited to gzip versions lower than 1.10.