In checks/changelog.pm, use centralized version parser, but in source facet.
The version parser in Lintian::Info::Changelog::Version splits version
strings into components, and checks the results for consistency. Uses
the centralized function when analyzing a changelog instead of relying
on custom regular expressions. It is more reliable.

The new parser code was adapted from the code that was here (removed
with this commit) with added modifications and improvements. One
important change is that, for native packages, the method->upstream
now always returns an empty string.

Also includes some cleanup regarding to processing order, variable
naming and simplified conditional logic.

Renames many variables to the 'latest' and 'previous' scheme. This was
super confusing before, when the 'first version' was actually the
latest one.

The email validation appeared incomplete: It seemed to ensure merely
that there was a dot in the domain. Now uses Email::Valid to ensure
addresses are acceptable.

Gbp-Dch: ignore