In Test::Lintian::Prepare, derive Prev-Version: from nativeness, not prior version.
Some tests examine what happens when versions change between native
and non-native. In some cases that causes Prev-Version: to have the
wrong format, too.

Uses the package's nativeness instead to decide on a format for the
calculated prior version (which may or may not be used by tests).

Gbp-Dch: ignore