Remove tag 'hyphen-in-native-debian-changelog-version'.
This tag can no longer be issued.  The new version parser raises an
exception before the tag can be issued.

At the same time, the centralized parser does not offer detailed
advice as to what is wrong, but it works better. The parser instead
issues the summary tag 'malformed-debian-changelog-version'.

Removes the tag hyphen-in-native-debian-changelog-version from
Lintian.

Gbp-Dch: ignore