In test 'changelog/changelog-file-missing-explicit-entry' for stable releases, use plus not tilde.
The tilde may be correct usage for some backports, but is super
confusing to the eye. When adjusting this test, the tag read:

    changelog-file-missing-explicit-entry (source): changelog-file-missing-explicit-entry 0.9-1 -> 1.2-3 (missing) -> 1.2-3~deb9u1

Debian's normal version scheme would not find that strange or
problematic.  Release candidates are commonly introduced via the
lower-ordering tilde.

Apparently, there is a special case for the specific suffix ~debXu1,
where X is a number.

Leaves the check alone, but modifies the test slightly so we do not
see the awkward condition again.

Gbp-Dch: ignore