Combine checks/source-copyright and checks/copyright-file into a new check 'copyright'.
The combination of these two related checks became possible by the
multi-facet design already used for changelogs, python, nodejs and
others.

The combined check is lengthy and will be partitioned again into
smaller and more manageable pieces, but the division will be topical
and input-invariant.

One advantage is that tags can be switched from binary to source, or
vice versa, with very little effort.

Gbp-Dch: ignore