In tests, add as expected copyright-related tags that appeared when two checks were consolidated.
The combination of the two checks copyright-file and source-copyright
into a single check copyright means that the combined check runs over
the combined tests. Some new tags appeared. They are added here.

This commit is part of an effort to combine to copyright-related
checks into a new check 'copyright'.  That was made possible by the
multi-facet feature for checks.

In the future, checks will primarily be topical and not so much
related to type of input.

Gbp-Dch: ignore