Combine two tests 'source-copyright-missing' from different checks into one.
These tests are very similar. Here, they are combined into a single
test with the same name. In this case, just the tags had to be
concatenated.

This commit is part of an effort to combine to copyright-related
checks into a new check 'copyright'.  That was made possible by the
multi-facet feature for checks.

In the future, checks will primarily be topical and not so much
related to type of input.

Gbp-Dch: ignore