Add new tag 'license-file-listed-in-debian-copyright' to avoid unnecessary clutter.
Currently, a tag unused-file-paragraph-in-dep5-copyright is issued
when a license file like COPYING or LICENSE is listed in d/copyright.
While that may be unnecessary, the tag is factually incorrect.  Also,
the severity minor indicates an error where there is none.

Provides the definition for a new tag that is to be issued instead in
the near future. The severity is wishlist.