Transfer binary heuristics for native status into checks/debian/changelog.pm.
The check debian/changelog depends in its binary segment on a
rudimentary form of native detection. The information is used to check
the installation paths for upstream and Debian changelogs, but not to
parse version numbers.

These checks cannot be be performed in source packages, and while they
should perhaps be part of the testsuite for dpkg-buildpackage instead,
it seemed unnecessary to lose this functionality right now.

Transfers the rudimentary native/non-native detection based on version
numbers out of Lintian::Collect::Binary, so it can be deleted from
there.

Gbp-Dch: ignore