In private/generate-profiles.pl, allot a full line for each value.
Shortens and simplifies the script that generates the profiles. The
output is identical with respect to format and content, but lists each
value on its own line. That way, it will be easier to see upcoming
changes in checks. The checks are about to be split up and sorted into
a hierarchy with subfolders.

Provides an updated script to generate profiles. The new profiles only
insert more whitespace.

Also imposes a an alphabetic sort order to fields in each paragraph.

Gbp-Dch: ignore