Regenerate profiles (no change).
These profiles remain identical with respect to content and format,
but all values are printed on their own lines.

It will make it easier to see the upcoming changes for checks. The
checks are about to be split into smaller parts that are sorted into a
hierarchy with subfolders.

The new profiles are also sorted alphabetically by field within each
paragraph.

Regenerates the Lintian profiles without functional changes.

Gbp-Dch: ignore