In tests, adjust expected literal tags for new check execution and tag emission order.
Except for the 'universal' output style, Lintian's output depends on
the order in which checks are run. This commit addresses the output
formats affected by the order of execution.

The checks are probably sorted alphabetically, and changing the check
watch-file to debian/watch means it gets executed earlier.

This is one of several upcoming changes that will help break down and
organize the massive and monolithic check scripts that accumulated
over the decades.

Gbp-Dch: ignore