Only complain about most recent historical changelog version; fix comparison without epochs.
The new version parser does not work on historical versions when the
nativeness has changed. It raises errors, so we cannot use it to strip
the epochs for all versions when looking for reused versions, which
was the only reason why historical versions were parsed.

This commit implements the epoch stripping with a regular expression
instead. It is no longer necessary to fully parse all historical
versions.

The most recent past version, i.e. the previous version, is still
parsed and may raise the tag odd-historical-debian-changelog-version.
That more or less replaces the functionality of the removed tag
latest-debian-changelog-entry-changed-to-native, except it makes no
claim about the nativeness of the previous upload. That cannot be
determined without querying the archive.

In some ways the new tag expands the functionality because it also
appears when the package switches from native to to non-native.

Gbp-Dch: ignore