In tests, adjust expected literal tags for new check execution and tag emission order.
Except for the 'universal' output style, Lintian's output depends on
the order in which checks are run. This commit addresses the output
formats affected by the order of execution.

The checks are probably sorted alphabetically, and changing a check
from checks/debian to checks/fields means it gets executed later.

Gbp-Dch: ignore