In Lintian::Collect, offer a method to access unfolded fields, in addition to literal fields.
Offers a new method 'unfolded_field' that operates mostly like the
existing method field, except fields are unfolded, i.e. newline
characters are removed.

Previously, the fields were unfolded in checks/fields.pm but that
check is in the process of being split up. The functionality is now
provided centrally.

The tag for multline-fields is being issued as it always has been,
except it will now be associated with the pseudo-check lintian.

Gbp-Dch: ignore