Fix command-line processing for --check-part so 'all' can also be part of a list.
The previous algorithm recognized 'all' only when it stood alone.
Recent changes in Lintian's internal architecture made it necessary at
times to add the pseudo-check 'lintian' to selected checks, which
could also happen in conjunction with 'all'. That did not work.

Now 'all' as an argument to '--check-part' can stand alone or be part
of a list.

Another subtle difference is that 'all' until now kept the enabled
profile unchanged, while going forward 'all' enables all known checks.
(That triggered a need to override the standard '--check-part all'
when testing with limited profiles such as ftp-master-rejects.)