As a default for tests, add pseudo-check 'lintian' when specifying a check to test.
Some tags are issued by Lintian proper and not by a particular check.
Those tags are associated with a pseudo-check called 'lintian'.  The
pseudo check cannot be tested on its own, as many processing requests
triggering the tags originate in "real" checks. The check 'lintian'
must be tested in conjunction with other checks.

As a default setting, adds the pseudo-check lintian when a check is
selected for testing. It happens often in the test suite.

Gbp-Dch: ignore