Fix check for license-file-listed-in-copyright; tag appeared too often.
The check issued the tag every time a license file was covered by
d/copyright, including through wildcards, which basically meant
always. It was much too broad.

Now the check identifies a list of license file among the source
files. For the tag to appear, the file must be listed explicitly, i.e.
not through a wildcard, in d/copyright.

Gbp-Dch: ignore