Update centralized version parser to include backports.
There are several ways to parse the backports. A decision was made to
incorporate them here. The advantage is that the check looks a lot
more logical, at least to the author.

At the same time, the additional complexity in the parser isn't great.
Perhaps the parsing could be structured better in the future by
wrapping the version components for versions without backports in
another Lintian::Info::Changelog::Version.

The point is to standardize version processing and hide the details.
Both should work to reduce potential bugs.

Gbp-Dch: ignore