Do not emit changelog-file-missing-explicit-entry when maintainer revision is 0. (Closes #940994)
Thanks to Simon McVittie for pointing it out, and for proposing a
great solution as well!

The tag was previously only issued for native packages, when it
related to NMUs (and not backports).  It is now also issued when NMUs
do not have a proper prior version for non-native packages.

The latter seemed like a reasonable enhancement of functionality. It
was not based on the bug report.