In Lintian::Collect, cache result of method unfolded_fields; prevents duplicate tags.
The method unfolded_field could previously not be called multiple
times because it can trigger a tag. That tag could then appear
multiple times.

This adds caching to the method. The tag is only emitted when the
value is added to the cache, i.e. in the first invocation. There no
longer a risk of duplicate tags.

The method can now be invoked multiple times in one check or even in
multiple checks.

Gbp-Dch: ignore