Split check/fields.pm into smaller components; invoke parts through old script.
This is the first step to divide the massive check 'fields' into
smaller components. They are invoked from the original script, meaning
they are not independent checks, in Lintian terminology.

Splitting the tests and the tags will take place later after it is
clear that no subchecks or tags need to be assigned to different
components. Then a '*.desc' file can be added, and the new check will
stand on its own.

This pattern may also be used for other large checks, such as files.