Remove support for native heuristics from Lintian facilities for installation (binary) packages.
There are many hints that a package is native, but the one that is
reliable only works for source packages. This removes from Lintian the
native detection facilities for installation (so-called binary)
packages.

There is history of inferring a native or non-native status for
installation (binary) packages from version numbers, but that
mechanism does not work all the time. There is also no reason why such
origin should matter for an installation package.

In Debian, a version string can only be parsed after the native or
non-native status was determined separately (which is no problem for
source packages). As a result of this requirement, many version
related tags can only issued for source packages.