Move check for tag 'rc-version-greater-than-expected-version' to check debian/changelog from fields/version.
This commit is a possible precursor to resolving Bug#782717. That fix
will require access to the previous changelog version, which is
available in the new location.

The mechanism to handle native package is slightly different because
the version parser in Lintian::Info::Changelog::Version returns native
package versions in the method maintainer_revision (but strips any
indicator for source NMUs).

The Dpkg parser, on the other hand, returned native package versions
in the same place as it would return an upstream version for a
non-native package.

Does not add back the regex string 'dev' that was removed in
Bug#732246.