Associate tags with new subchecks for original check 'fields'.
In the near future, the subchecks of 'fields' will stand on their own.
They will no longer be invoked from inside the old check.

This associates the tags emitted by each subcheck with that subcheck.

The split was easy except for maintainer and uploaders, which are
somewhat awkwardly issued by a library module. (There are many
similarities in the way they are processed.) A better way would be for
the library to evaluate the fields, but for the checks to issue their
own tags.