Split tests and sort them into the appropriate subchecks for the old check 'fields'.
Tests were copied or duplicated as needed whenever it helped to
evaluate a subcheck. The two scenarios were either that the expected
tags or the tags listed in Test-Against related to the subcheck.

One exception was the test fields-depends-wildcard, which had neither
expected any tags nor defined a Test-Against. Upon examination, that
test was added to the check 'fields/package-relations' with
appropriate settings in Test-Against.

Splits all tests for the old check 'fields' into the subtests so they
can stand on their own.

Gbp-Dch: ignore