In tests, load profiles only from one location; make sure one is loaded in pod-syposis.t. (May fix #942252)
Some test scripts appear to load old profiles. They are presumably
located in the installed location /usr/share/lintian. The errors
depend on the version difference between what is installed and what is
under test. That is probably why the reported bug does not show up in
Gitlab (anymore).

This commit does two things: (1) It unambiguously uses only one
include path when finding profiles, and (2) makes sure a profile is
loaded in the test pod-synopsis.t that fails in the bug report.

This commit is an attempt to resolve Bug#942252. It causes no
breakage, but the bug did never appeared locally. If it works, the bug
report should be closed manually.